backend/requirements.txt
Flask==2.0.1
requests==2.26.0
psycopg2==2.9.1
schedule==1.1.0
PyYAML==5.4.1